fixing promotor suggestions, deleting id from ticket and application, changing dashbord view to only open jobs

+++ src/main/java/pl/wroblewski/helpdeskapp/models/Application.java
@@ -26,7 +26,7 @@ public class Application {
     private String typeOfApplication;
 
     @Column(name = "application_number")
-    private Long applicationNumber;
+    private Integer applicationNumber;
 
     @Column(name = "description", length = 300)
     private String description;
